refactor(blog): extract PostArticleBody and post shared helpers
Share sidebar, prose, footer, and comment chrome between PostLayout and
the upcoming spotlight layout.
